Huge Update: Cursor Control added, for other changes, see description
The option to control Player 1s movement (not player 2, as in multiplayer, the cursor would have to be constantly moving, and the player would stop on touching it) with the cursor / mouse is added in the options scene, and instructions scene. instructions scene updated to include this, and fix a wording mistake, as well as a small design mistake. speed decreased by 40% in response to feedback, and to keep the game fair, it also takes 40% longer time for food to decrease. Finally, the process of calculating the score has changed, and is now more accurate, and exponential.